Refactor to pull TreeNode class into its own file tree_node.py
